Stable run for Cluster Torture tests
